Start running server, Keep in mind to handle server/agent mode in all code
