+++ src/log.hpp
@@ -6,14 +6,12 @@
 #include <ctime>
 #include <iomanip>
 
-#include "config.hpp"
-
 class Log
 {
 
 private:
 		
-	const Config config_;
+	const Config& config_;
 	bool logFileClosed;
 
 	/*
@@ -71,12 +69,48 @@ public:
 	/*
 	*	This fucntion returns string of Log::Level
 	*/
-	const std::string levelToString(const Level level);
+	const std::string levelToString(const Level level)
+	{
+		std::string result("");
+		switch (level){
+			case Level::INFO:
+				result = "INFO";
+				break;
+			case Level::ERROR:
+				result = "ERROR";
+				break;
+			case Level::DEBUG:
+				result = "DEBUG";
+				break;
+		}
+		return result;
+	}
 
 	/*
 	*	This function writes message in to the log file
 	*/
-	void write(const std::string message, const Level level);
+	void write(const std::string message, const Level level)
+	{
+		std::ofstream logFile_(config_.log_.file, logFile_.out | logFile_.app);
+		if (level <= level_ || level == Level::ERROR){
+			if (logFile_.is_open())
+			{
+				time_t now = time(0);
+				auto time = std::time(nullptr);
+				auto localtime = *std::localtime(&time);
+				std::ostringstream oss;
+				oss << std::put_time(&localtime, "%Y-%m-%d_%H:%M:%S");
+				auto timeStr = oss.str();
+				std::string line = timeStr + " [" + levelToString(level) + "] " + message + "\n";
+				logFile_ << line;
+				std::cout << line;
+				logFile_.close();
+			} else 
+			{
+				std::cerr << "Error openning log file : " << config_.log_.file  << " make sure the directory and file exists " << "\n";
+			}
+		}
+	}
 	
 };
