+++ src/session.hpp
@@ -1,5 +1,12 @@
 #ifndef SESSION_HPP
 #define SESSION_HPP
 
+class Session
+{
+public:
+	Session();
+	~Session();
+	
+};
 
 #endif /* SESSION_HPP */
